requirements.txt
# To install: pip install -r requirements.txt

torch==2.0.1
torchvision==0.15.2
timm==0.9.2
albumentations==1.3.0
scikit-learn==1.2.2
numpy==1.24.3
tqdm==4.65.0
pillow==9.5.0
matplotlib==3.7.1
faiss-cpu==1.7.4
torchviz==0.0.2
